visualize a tree

plots a tree in the forest.

Usage

# S3 method for forestry
plot(x, tree.id = 1, print.meta_dta = FALSE, beta.char.len = 30, ...)

Arguments

x

A forestry x.

tree.id

Specifies the tree number that should be visualized.

print.meta_dta

A flag indicating whether the data for the plot should be printed.

beta.char.len

The length of the beta values in leaf node representation. This is only required when plotting a forestry object with linear aggregation functions (linear = TRUE).

...

additional arguments that are not used.

Value

A plot of the specified tree in the forest.

Examples

set.seed(292315)
rf <- forestry(x = iris[,-1],
               y = iris[, 1],
               nthread = 2)

plot(x = rf)

plot(x = rf, tree.id = 2)

plot(x = rf, tree.id = 500)